* extract JInput helper code * downgrade version of JInput version to 2.0.5
