+++ pom.xml
@@ -12,7 +12,7 @@
         <dependency>
             <groupId>net.java.jinput</groupId>
             <artifactId>jinput</artifactId>
-            <version>2.0.6</version>
+            <version>2.0.5</version>
         </dependency>
     </dependencies>
     <properties>
